Milestone bugfix
This was a prime example of murphy's law, I experienced all my
cleverness go up in smoke when I tried to do a "hot" deployment
of the systemd services. Lesson learnt, always assume that everything
shall go to shit. I discovered some bugs too on some old sections in
box_admin. Bug fixing is a continuous process and you can never forsee
all the possibilities. It is a valuable thing to get to finally learn.