Sqlbox systemd service experiment
To make sure that systemd is informed of a box's death have the
startup script check the life of the boxes it has started and on
detecting any "death" inform systemd that a restart is warranted.

The service definition then changes to reflect this also adding an
ExecStartPre directive to kill any running boxes before attempting
to start them.

box_start has also been modified to return the name of the pidfile
of the started box.