Bugfix on logrotate config infection and file names.
Aside from the gateway folder on the logrotate directory being
incorrectly specified, the access log has now been moved into its
own directory under the $LOGDIR and necessary adjustments made in
the config tarball and the logrotate config file.

The placeholder gateway name in the logrotate config file is "testing"
just as it is in the config tarball.